pyproject.toml
[project]
name = "mia-vpc3"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "albumentations>=2.0.8",
    "matplotlib>=3.10.0",
    "pyyaml>=6.0.3",
    "torchmetrics>=1.9.0",
    "torchvision>=0.26.0",
    "transformers>=5.3.0",
    "wandb>=0.25.1",
]
